Move make_contractions from gbasis.contractions to gbasis.parsers
Since make_contractions method is specific to the way that the parsers store the
information, we move it to the gbasis.parsers module to make sure that (1) the
convention for storage can be found within the module and (2) changes to either
the parser or the make_contractions is localized to one module.

1. Move make_contractions function from contractions module to parsers module
2. Update tests
3. Update docstring in contractions
